chore(elasticsearch sink): split retry module
