+++ src/sinks/elasticsearch/mod.rs
@@ -1,21 +1,21 @@
+mod retry;
+
 use crate::{
     config::{DataType, SinkConfig, SinkContext, SinkDescription},
     emit,
     event::{Event, Value},
-    http::{Auth, HttpClient, HttpError, MaybeAuth},
+    http::{Auth, HttpClient, MaybeAuth},
     internal_events::{ElasticSearchEventEncoded, TemplateRenderingFailed},
     rusoto::{self, region_from_endpoint, AwsAuthentication, RegionOrEndpoint},
     sinks::util::{
         encoding::{EncodingConfigWithDefault, EncodingConfiguration},
         http::{BatchedHttpSink, HttpSink, RequestConfig},
-        retries::{RetryAction, RetryLogic},
         BatchConfig, BatchSettings, Buffer, Compression, EncodedEvent, TowerRequestConfig,
         UriSerde,
     },
     template::{Template, TemplateParseError},
     tls::{TlsOptions, TlsSettings},
 };
-use bytes::Bytes;
 use futures::{FutureExt, SinkExt};
 use http::{
     header::{HeaderName, HeaderValue},
@@ -25,6 +25,7 @@ use http::{
 use hyper::Body;
 use indexmap::IndexMap;
 use lazy_static::lazy_static;
+use retry::ElasticSearchRetryLogic;
 use rusoto_core::Region;
 use rusoto_credential::{CredentialsError, ProvideAwsCredentials};
 use rusoto_signature::{SignedRequest, SignedRequestPayload};
@@ -597,92 +598,6 @@ impl HttpSink for ElasticSearchCommon {
     }
 }
 
-#[derive(Clone)]
-struct ElasticSearchRetryLogic;
-
-#[derive(Deserialize, Debug)]
-struct EsResultResponse {
-    items: Vec<EsResultItem>,
-}
-#[derive(Deserialize, Debug)]
-enum EsResultItem {
-    #[serde(rename = "index")]
-    Index(EsIndexResult),
-    #[serde(rename = "create")]
-    Create(EsIndexResult),
-}
-#[derive(Deserialize, Debug)]
-struct EsIndexResult {
-    error: Option<EsErrorDetails>,
-}
-#[derive(Deserialize, Debug)]
-struct EsErrorDetails {
-    reason: String,
-    #[serde(rename = "type")]
-    err_type: String,
-}
-
-impl EsResultItem {
-    fn result(self) -> EsIndexResult {
-        match self {
-            EsResultItem::Index(r) => r,
-            EsResultItem::Create(r) => r,
-        }
-    }
-}
-
-impl RetryLogic for ElasticSearchRetryLogic {
-    type Error = HttpError;
-    type Response = hyper::Response<Bytes>;
-
-    fn is_retriable_error(&self, _error: &Self::Error) -> bool {
-        true
-    }
-
-    fn should_retry_response(&self, response: &Self::Response) -> RetryAction {
-        let status = response.status();
-
-        match status {
-            StatusCode::TOO_MANY_REQUESTS => RetryAction::Retry("too many requests".into()),
-            StatusCode::NOT_IMPLEMENTED => {
-                RetryAction::DontRetry("endpoint not implemented".into())
-            }
-            _ if status.is_server_error() => RetryAction::Retry(format!(
-                "{}: {}",
-                status,
-                String::from_utf8_lossy(response.body())
-            )),
-            _ if status.is_client_error() => {
-                let body = String::from_utf8_lossy(response.body());
-                RetryAction::DontRetry(format!("client-side error, {}: {}", status, body))
-            }
-            _ if status.is_success() => {
-                let body = String::from_utf8_lossy(response.body());
-
-                if body.contains("\"errors\":true") {
-                    RetryAction::DontRetry(get_error_reason(&body))
-                } else {
-                    RetryAction::Successful
-                }
-            }
-            _ => RetryAction::DontRetry(format!("response status: {}", status)),
-        }
-    }
-}
-
-fn get_error_reason(body: &str) -> String {
-    match serde_json::from_str::<EsResultResponse>(&body) {
-        Err(json_error) => format!(
-            "some messages failed, could not parse response, error: {}",
-            json_error
-        ),
-        Ok(resp) => match resp.items.into_iter().find_map(|item| item.result().error) {
-            Some(error) => format!("error type: {}, reason: {}", error.err_type, error.reason),
-            None => format!("error response: {}", body),
-        },
-    }
-}
-
 impl ElasticSearchCommon {
     pub fn parse_config(config: &ElasticSearchConfig) -> crate::Result<Self> {
         // Test the configured host, but ignore the result
@@ -838,8 +753,7 @@ fn maybe_set_id(key: Option<impl AsRef<str>>, doc: &mut serde_json::Value, event
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{event::Event, event::Value, sinks::util::retries::RetryAction};
-    use http::{Response, StatusCode};
+    use crate::event::{Event, Value};
     use pretty_assertions::assert_eq;
     use serde_json::json;
     use std::collections::BTreeMap;
@@ -1073,34 +987,6 @@ mod tests {
 {"data_stream":{"dataset":"testing","type":"synthetics"},"message":"hello there","timestamp":"2020-12-01T01:02:03Z"}
 "#;
         assert_eq!(std::str::from_utf8(&encoded).unwrap(), expected);
-    }
-
-    #[test]
-    fn handles_error_response() {
-        let json = "{\"took\":185,\"errors\":true,\"items\":[{\"index\":{\"_index\":\"test-hgw28jv10u\",\"_type\":\"log_lines\",\"_id\":\"3GhQLXEBE62DvOOUKdFH\",\"status\":400,\"error\":{\"type\":\"illegal_argument_exception\",\"reason\":\"mapper [message] of different type, current_type [long], merged_type [text]\"}}}]}";
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .body(Bytes::from(json))
-            .unwrap();
-        let logic = ElasticSearchRetryLogic;
-        assert!(matches!(
-            logic.should_retry_response(&response),
-            RetryAction::DontRetry(_)
-        ));
-    }
-
-    #[test]
-    fn get_index_error_reason() {
-        let json = "{\"took\":185,\"errors\":true,\"items\":[{\"index\":{\"_index\":\"test-hgw28jv10u\",\"_type\":\"log_lines\",\"_id\":\"3GhQLXEBE62DvOOUKdFH\",\"status\":400,\"error\":{\"type\":\"illegal_argument_exception\",\"reason\":\"mapper [message] of different type, current_type [long], merged_type [text]\"}}}]}";
-        let reason = get_error_reason(&json);
-        assert_eq!(reason, "error type: illegal_argument_exception, reason: mapper [message] of different type, current_type [long], merged_type [text]");
-    }
-
-    #[test]
-    fn get_create_error_reason() {
-        let json = "{\"took\":3,\"errors\":true,\"items\":[{\"create\":{\"_index\":\"test-hgw28jv10u\",\"_type\":\"_doc\",\"_id\":\"aBLq1HcBWD7eBWkW2nj4\",\"status\":400,\"error\":{\"type\":\"mapper_parsing_exception\",\"reason\":\"object mapping for [host] tried to parse field [host] as object, but found a concrete value\"}}}]}";
-        let reason = get_error_reason(&json);
-        assert_eq!(reason, "error type: mapper_parsing_exception, reason: object mapping for [host] tried to parse field [host] as object, but found a concrete value");
     }
 
     #[test]
